Add category backups management functionality
Add functionality to this plugin, which allows backups management on a category context to all users with the appropriate capabilities.
This commit also adds an additional plugin settings page with a database query optimization option, which greatly reduces the amount
of time it takes for the query to execute, but decreases accuracy of finding the backup files to only those backups, which have been created
on the current Moodle instance (no user uploaded files). The increased query execution speed helps on large Moodle installations.
Also added slovenian translations.

+++ categorybackups.php
@@ -23,23 +23,66 @@
  */
 
 require_once('../../config.php');
-require_once($CFG->libdir . '/adminlib.php');
+require_login();
 
+require_once($CFG->libdir.'/filelib.php');
+require_once($CFG->libdir . '/accesslib.php');
+
+global $PAGE;
+
+$contextid = intval(required_param('contextid', PARAM_INT));
 $delete = optional_param('delete', '', PARAM_TEXT);
 $filename = optional_param('filename', '', PARAM_TEXT);
 $deleteselected = optional_param('deleteselectedfiles', '', PARAM_TEXT);
 $fileids = optional_param('fileids', '', PARAM_TEXT);
 $currenttab = optional_param('tab', 'core', PARAM_TEXT);
 
-admin_externalpage_setup('reportallbackups_system', '', array('tab' => $currenttab), '', array('pagelayout' => 'report'));
+$urlparams = array("contextid" => $contextid, "tab" => $currenttab);
+if (!empty($delete)) {
+    $urlparams['delete'] = $delete;
+}
+if (!empty($deleteselected)) {
+    $urlparams['deleteselectedfiles'] = $deleteselected;
+}
+if (!empty($filename)) {
+    $urlparams['filename'] = $filename;
+}
+if (!empty($fileids)) {
+    $urlparams['fileids'] = $fileids;
+}
+$PAGE->set_url(new moodle_url('/report/allbackups/categorybackups.php', $urlparams));
+// Set page context as category context, defined by the required param $contextid.
+$context = context::instance_by_id($contextid);
+$PAGE->set_context($context);
+// Set this page as an instance of a category.
+$PAGE->set_category_by_id($context->instanceid);
+// Set page heading and title.
+$PAGE->set_title(get_string('pluginname', 'report_allbackups'));
+$PAGE->set_heading(get_string('pluginname', 'report_allbackups'));
+// Set page layout as admin.
+$PAGE->set_pagelayout('report');
+// Add navigation link to the category this page belongs to.
+$PAGE->navbar->ignore_active();
+$PAGE->navbar->add($context->get_context_name(false), new moodle_url('/course/index.php', array(
+    'categoryid' => $context->instanceid
+)));
+// Check plugin config for category backup management mode.
+$pluginconfig = get_config('report_allbackups');
+if (!$pluginconfig->categorybackupmgmt) {
+    echo $OUTPUT->header();
+    \core\notification::error(get_string('error:categorybackupmgmtmodedisabled', 'report_allbackups'));
+    echo $OUTPUT->footer();
+    exit;
+}
+
+require_capability('report/categorybackups:view', $PAGE->context, $USER->id);
 
 $backupdest = get_config('backup', 'backup_auto_destination');
 if (empty($backupdest) && $currenttab == 'autobackup') {
     print_error(get_string("autobackupnotset", "report_allbackups"));
 }
 
-$context = context_system::instance();
-if (has_capability('report/allbackups:delete', $context)) {
+if (has_capability('report/categorybackups:delete', $PAGE->context)) {
     if (!empty($deleteselected) || !empty($delete)) { // Delete action.
         if (empty($fileids)) {
             $fileids = array();
@@ -66,11 +109,16 @@ if (has_capability('report/allbackups:delete', $context)) {
             }
             // Display confirmation box - are you really sure you want to delete this file?
             echo $OUTPUT->header();
-            $params = array('deleteselectedfiles' => 1, 'confirm' => 1, 'fileids' => implode(',', $fileids), 'tab' => $currenttab);
+            $params = array(
+                'deleteselectedfiles' => 1,
+                'confirm' => 1,
+                'fileids' => implode(',', $fileids),
+                'tab' => $currenttab,
+                'contextid' => $PAGE->context->id);
             $deleteurl = new moodle_url($PAGE->url, $params);
             $numfiles = count($fileids);
             echo $OUTPUT->confirm(get_string('areyousurebulk', 'report_allbackups', $numfiles),
-                $deleteurl, $CFG->wwwroot . '/report/allbackups/index.php');
+                $deleteurl, $CFG->wwwroot . '/report/allbackups/categorybackups.php?contextid='.$PAGE->context->id);
 
             echo $OUTPUT->footer();
             exit;
@@ -82,12 +130,13 @@ if (has_capability('report/allbackups:delete', $context)) {
                     // Check nothing weird passed in filename - protect against directory traversal etc.
                     // Check to make sure this is an mbz file.
                     if ($id == clean_param($id, PARAM_FILE) &&
-                        pathinfo($id, PATHINFO_EXTENSION) == 'mbz' &&
-                        is_readable($backupdest .'/'. $id)) {
+                            pathinfo($id, PATHINFO_EXTENSION) == 'mbz' &&
+                            is_readable($backupdest .'/'. $id) &&
+                            has_capability('moodle/course:delete', $PAGE->context)) {
 
                         unlink($backupdest .'/'. $id);
                         $event = \report_allbackups\event\autobackup_deleted::create(array(
-                            'context' => context_system::instance(),
+                            'context' => $PAGE->context,
                             'objectid' => null,
                             'other' => array('filename' => $id)));
                         $event->trigger();
@@ -100,7 +149,9 @@ if (has_capability('report/allbackups:delete', $context)) {
                     $file = $fs->get_file_by_id((int)$id);
                     $fileext = pathinfo($file->get_filename(), PATHINFO_EXTENSION);
                     // Make sure the file exists, and it is a backup file we are deleting.
-                    if (!empty($file) && $fileext == 'mbz') {
+                    if (!empty($file) &&
+                            $fileext == 'mbz' &&
+                            has_capability('moodle/course:delete', context::instance_by_id($file->get_contextid()))) {
                         $file->delete();
                         $event = \report_allbackups\event\backup_deleted::create(array(
                             'context' => context::instance_by_id($file->get_contextid()),
@@ -124,13 +175,13 @@ if ($currenttab == 'autobackup') {
     $filters = array('filename' => 0, 'realname' => 0, 'coursecategory' => 0, 'filearea' => 0, 'timecreated' => 0);
 }
 if ($currenttab == 'autobackup') {
-    $table = new \report_allbackups\output\autobackups_table('autobackups');
+    $table = new \report_allbackups\output\autobackups_table('autobackups', $PAGE->context);
 } else {
-    $table = new \report_allbackups\output\allbackups_table('allbackups');
+    $table = new \report_allbackups\output\allbackups_table('allbackups', $PAGE->context);
     $table->define_baseurl($PAGE->url);
 }
 
-$ufiltering = new \report_allbackups\output\filtering($filters, $PAGE->url);
+$ufiltering = new \report_allbackups\output\filtering($filters, $PAGE->url, null, true);
 if (!$table->is_downloading()) {
     // Only print headers if not asked to download data.
     // Print the page header.
@@ -139,10 +190,10 @@ if (!$table->is_downloading()) {
     if (!empty(get_config('backup', 'backup_auto_destination'))) {
         $row = $tabs = array();
         $row[] = new tabobject('core',
-            $CFG->wwwroot.'/report/allbackups',
+            $CFG->wwwroot.'/report/allbackups/categorybackups.php?contextid='.$PAGE->context->id,
             get_string('standardbackups', 'report_allbackups'));
         $row[] = new tabobject('autobackup',
-            $CFG->wwwroot.'/report/allbackups/index.php?tab=autobackup',
+            $CFG->wwwroot.'/report/allbackups/categorybackups.php?contextid='.$PAGE->context->id.'&tab=autobackup',
             get_string('autobackup', 'report_allbackups'));
         $tabs[] = $row;
         print_tabs($tabs, $currenttab);
@@ -155,7 +206,7 @@ if (!$table->is_downloading()) {
     $ufiltering->display_add();
     $ufiltering->display_active();
 
-    echo '<form action="index.php" method="post" id="allbackupsform">';
+    echo '<form action="categorybackups.php?contextid='.$PAGE->context->id.'" method="post" id="allbackupsform">';
     echo html_writer::start_div();
     echo html_writer::tag('input', '', array('type' => 'hidden', 'name' => 'sesskey', 'value' => sesskey()));
     echo html_writer::tag('input', '', array('type' => 'hidden', 'name' => 'returnto', 'value' => s($PAGE->url->out(false))));
@@ -167,7 +218,7 @@ if (!$table->is_downloading()) {
 }
 if ($currenttab == 'autobackup') {
     // Get list of files from backup.
-    $table->adddata($ufiltering);
+    $table->adddata($ufiltering, $PAGE->context);
 } else {
     list($extrasql, $params) = $ufiltering->get_sql_filter();
     $fields = "f.id, f.contextid, f.component, f.filearea, f.filename, f.userid, f.filesize, f.timecreated, f.filepath, f.itemid, ";
@@ -180,14 +231,16 @@ if ($currenttab == 'autobackup') {
             $from = "(
                 SELECT *
                 FROM {files}
-                WHERE component=:cmpbackup OR (component=:cmpcourse AND filearea=:falegacy)
+                WHERE component=:cmpbackup
+                    OR (component=:cmpcourse AND filearea=:falegacy)
                 ) f
                 JOIN {user} u on u.id = f.userid";
         } else {
             $from = "(
                 SELECT *
                 FROM {files}
-                WHERE filearea IN (:facourse, :faautomated, :falegacy) AND component IN (:cmpbackup, :cmpcourse)
+                WHERE filearea IN (:facourse, :faautomated, :falegacy)
+                    AND component IN (:cmpbackup, :cmpcourse)
                 ) f
                 JOIN {user} u on u.id = f.userid";
 
@@ -202,7 +255,7 @@ if ($currenttab == 'autobackup') {
             $from .= " JOIN {course} c ON c.id = cx.instanceid";
         }
 
-        $where = "f.filename LIKE :backupsuffix";
+        $where = "f.filename LIKE :backupsuffix AND cx.path LIKE '".$context->path."%'";
 
         $params['falegacy'] = 'legacy';
         $params['cmpbackup'] = 'backup';
@@ -211,13 +264,18 @@ if ($currenttab == 'autobackup') {
         $params['backupsuffix'] = '%.mbz';
 
     } else {
-        $from = "{files} f JOIN {user} u on u.id = f.userid";
+        $from = "{files} f
+            JOIN {user} u on u.id = f.userid
+            JOIN {context} cx ON cx.id = f.contextid
+                AND cx.contextlevel = :contextlevel";
         if (strpos($extrasql, 'c.category') !== false) {
             // Category filter included, Join with course table.
-            $from .= " JOIN {context} cx ON cx.id = f.contextid AND cx.contextlevel = :contextlevel
-                JOIN {course} c ON c.id = cx.instanceid";
+            $from .= " JOIN {course} c ON c.id = cx.instanceid";
         }
-        $where = "f.filename LIKE :backupsuffix AND f.component <> :cmprecycle AND f.filearea <> :fadraft";
+        $where = "f.filename LIKE :backupsuffix
+            AND f.component <> :cmprecycle
+            AND f.filearea <> :fadraft
+            AND cx.path LIKE '".$context->path."%'";
 
         $params['cmprecycle'] = 'tool_recyclebin';
         $params['fadraft'] = 'draft';
@@ -233,7 +291,7 @@ if ($currenttab == 'autobackup') {
     $table->out(40, true);
 }
 
-if (!$table->is_downloading()) {
+if (!$table->is_downloading() && has_capability('moodle/course:delete', $PAGE->context)) {
     echo html_writer::tag('input', "", array('name' => 'deleteselectedfiles', 'type' => 'submit',
         'id' => 'deleteallselected', 'class' => 'btn btn-secondary',
         'value' => get_string('deleteselectedfiles', 'report_allbackups')));
